minor update to random strategy

+++ src/TestingServices/ExplorationStrategies/RandomStrategy.cpp
@@ -21,9 +21,9 @@ TestingServices::RandomStrategy::RandomStrategy()
 {
     std::random_device rd;
     std::mt19937 gen(rd());
-    std::uniform_int_distribution<unsigned int> dis(std::mt19937::min(), std::mt19937::max());
+    std::uniform_int_distribution<size_t> dis(std::mt19937::min(), std::mt19937::max());
     m_seed = dis(gen);
-    m_generator.seed(m_seed);
+    m_generator.seed((unsigned int)m_seed);
 }
 
 bool TestingServices::RandomStrategy::TryGetNext(ActorInfo*& next,
@@ -43,7 +43,7 @@ bool TestingServices::RandomStrategy::TryGetNext(ActorInfo*& next,
         return false;
     }
 
-    std::uniform_int_distribution<unsigned int> dis(0, enabled.size() - 1);
+    std::uniform_int_distribution<size_t> dis(0, enabled.size() - 1);
     next = enabled[dis(m_generator)];
 
     return true;
