+++ src/TestingServices/ExplorationStrategies/RandomStrategy.h
@@ -36,9 +36,8 @@ namespace Microsoft { namespace P3 { namespace TestingServices
         bool GetNextBooleanChoice(int maxValue, bool& next);
             
     private:
-
         // Seed used during this iteration.
-        unsigned int m_seed;
+        size_t m_seed;
 
         // Random integer generator.
         std::mt19937 m_generator;
